+++ src/torchoutil/utils/saving/__init__.py
@@ -4,7 +4,7 @@
 from torchoutil.utils.packaging import _YAML_AVAILABLE
 
 from .common import to_builtin
-from .csv_io import load_csv, save_to_csv
+from .csv_io import load_csv, to_csv
 
 if _YAML_AVAILABLE:
-    from .yaml_io import load_yaml, save_to_yaml
+    from .yaml_io import load_yaml, to_yaml